nest tools under tools
